Move _verify_source_file_access to live with similar functions
